Update service to Python 3.11 and earthdata-varinfo. Update notebook dependencies.
Squashed commit of the following:

commit 6b66a2592458df96ddc45bdba7c24e50927ff1d4
Author: Owen Littlejohns <owen.m.littlejohns@nasa.gov>
Date:   Thu Jul 27 12:41:10 2023 -0400

    Update recommended notebook Python version to 3.11

commit 07f59b596c3357c65be2d23f24c4ea12e74e1737
Author: Owen Littlejohns <owen.m.littlejohns@nasa.gov>
Date:   Thu Jul 20 22:57:38 2023 -0400

    Update service dependencies and Python environment to 3.11.

commit 6edde6dd7d4508d60c86645dca4a7f8c62666fcb
Author: Owen Littlejohns <owen.m.littlejohns@nasa.gov>
Date:   Wed Apr 5 20:59:22 2023 -0400

    Update documentation notebook dependencies to mitigate Snyk vulnerability.

+++ test/unit/test_swotrepr_geometry.py
@@ -43,10 +43,10 @@ class TestSwotReprGeometry(TestBase):
         cls.test_file = Dataset(cls.test_path, 'w')
         cls.test_file.createDimension('nj', size=3)
         cls.test_file.createDimension('ni', size=4)
-        cls.test_file.createVariable('lat', np.float, dimensions=('nj', 'ni'))
-        cls.test_file.createVariable('lon', np.float, dimensions=('nj', 'ni'))
-        cls.test_file.createVariable('lat_1d', np.float, dimensions=('ni',))
-        cls.test_file.createVariable('lon_1d', np.float, dimensions=('ni',))
+        cls.test_file.createVariable('lat', float, dimensions=('nj', 'ni'))
+        cls.test_file.createVariable('lon', float, dimensions=('nj', 'ni'))
+        cls.test_file.createVariable('lat_1d', float, dimensions=('ni',))
+        cls.test_file.createVariable('lon_1d', float, dimensions=('ni',))
         cls.test_file['lat'][:] = cls.lat_data
         cls.test_file['lon'][:] = cls.lon_data
         cls.test_file['lat_1d'][:] = np.array([0.0, 3.0, 6.0, 9.0])
@@ -286,9 +286,9 @@ class TestSwotReprGeometry(TestBase):
                 test_file = Dataset(test_path, 'w')
                 test_file.createDimension('nj', size=2)
                 test_file.createDimension('ni', size=2)
-                test_file.createVariable('lat', np.float, dimensions=('nj', 'ni'),
+                test_file.createVariable('lat', float, dimensions=('nj', 'ni'),
                                          fill_value=fill_value)
-                test_file.createVariable('lon', np.float, dimensions=('nj', 'ni'),
+                test_file.createVariable('lon', float, dimensions=('nj', 'ni'),
                                          fill_value=fill_value)
                 test_file['lat'][:] = lat_data
                 test_file['lon'][:] = lon_data
